requirements.txt
dash==2.13.0
dash_table==5.0.0
mysql_connector_repackaged==0.3.1
pandas==2.1.0
plotly==5.17.0
scikit_learn==1.3.0
SQLAlchemy==2.0.20
